refactor and Reputation type
